better neovim install

+++ install.sh
@@ -33,12 +33,15 @@ rm -rf "${HOME}/.config/nvim"
 cp -r "${HOME}"/.dotfiles/nvim "${HOME}"/.config/nvim
 
 # install neovim
-if [[ "$OSTYPE" == "linux-gnu"* ]]; then
-	curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.tar.gz
-	sudo rm -rf /opt/nvim-linux-x86_64
-	sudo tar -C /opt -xzf nvim-linux-x86_64.tar.gz
-elif [[ "$OSTYPE" == "darwin"* ]]; then
-	curl -LO https://github.com/neovim/neovim/releases/download/nightly/nvim-macos-x86_64.tar.gz
-	tar xzf nvim-macos-x86_64.tar.gz
-	./nvim-macos-x86_64/bin/nvim
+if command -v nvim &> /dev/null
+then
+    if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+        curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.tar.gz
+        sudo rm -rf /opt/nvim-linux-x86_64
+        sudo tar -C /opt -xzf nvim-linux-x86_64.tar.gz
+    elif [[ "$OSTYPE" == "darwin"* ]]; then
+        curl -LO https://github.com/neovim/neovim/releases/download/nightly/nvim-macos-x86_64.tar.gz
+        tar xzf nvim-macos-x86_64.tar.gz
+        ./nvim-macos-x86_64/bin/nvim
+    fi
 fi
